Restructure packaging for blik
COAND-497
