+++ blik/src/main/java/com/adyen/checkout/blik/BlikComponent.kt
@@ -14,6 +14,8 @@ import com.adyen.checkout.action.ActionHandlingComponent
 import com.adyen.checkout.action.DefaultActionHandlingComponent
 import com.adyen.checkout.action.GenericActionDelegate
 import com.adyen.checkout.blik.BlikComponent.Companion.PROVIDER
+import com.adyen.checkout.blik.internal.provider.BlikComponentProvider
+import com.adyen.checkout.blik.internal.ui.BlikDelegate
 import com.adyen.checkout.components.ButtonComponent
 import com.adyen.checkout.components.PaymentComponent
 import com.adyen.checkout.components.PaymentComponentEvent
