+++ blik/src/main/java/com/adyen/checkout/blik/internal/ui/BlikDelegate.kt
@@ -6,8 +6,10 @@
  * Created by josephj on 1/7/2022.
  */
 
-package com.adyen.checkout.blik
+package com.adyen.checkout.blik.internal.ui
 
+import com.adyen.checkout.blik.internal.ui.model.BlikInputData
+import com.adyen.checkout.blik.internal.ui.model.BlikOutputData
 import com.adyen.checkout.components.PaymentComponentState
 import com.adyen.checkout.components.base.PaymentComponentDelegate
 import com.adyen.checkout.components.model.payments.request.BlikPaymentMethod
@@ -16,7 +18,7 @@ import com.adyen.checkout.components.ui.UIStateDelegate
 import com.adyen.checkout.components.ui.ViewProvidingDelegate
 import kotlinx.coroutines.flow.Flow
 
-interface BlikDelegate :
+internal interface BlikDelegate :
     PaymentComponentDelegate<PaymentComponentState<BlikPaymentMethod>>,
     ViewProvidingDelegate,
     ButtonDelegate,
