+++ blik/src/main/java/com/adyen/checkout/blik/internal/ui/DefaultBlikDelegate.kt
@@ -6,14 +6,17 @@
  * Created by josephj on 1/7/2022.
  */
 
-package com.adyen.checkout.blik
+package com.adyen.checkout.blik.internal.ui
 
+import androidx.annotation.VisibleForTesting
 import androidx.lifecycle.LifecycleOwner
+import com.adyen.checkout.blik.internal.ui.model.BlikInputData
+import com.adyen.checkout.blik.internal.ui.model.BlikOutputData
 import com.adyen.checkout.components.PaymentComponentEvent
 import com.adyen.checkout.components.PaymentComponentState
 import com.adyen.checkout.components.analytics.AnalyticsRepository
 import com.adyen.checkout.components.base.ButtonComponentParams
-import com.adyen.checkout.components.model.paymentmethods.StoredPaymentMethod
+import com.adyen.checkout.components.model.paymentmethods.PaymentMethod
 import com.adyen.checkout.components.model.payments.request.BlikPaymentMethod
 import com.adyen.checkout.components.model.payments.request.OrderRequest
 import com.adyen.checkout.components.model.payments.request.PaymentComponentData
@@ -32,14 +35,16 @@ import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.launch
 
 @Suppress("TooManyFunctions")
-internal class StoredBlikDelegate(
+internal class DefaultBlikDelegate(
     private val observerRepository: PaymentObserverRepository,
     override val componentParams: ButtonComponentParams,
-    private val storedPaymentMethod: StoredPaymentMethod,
+    private val paymentMethod: PaymentMethod,
     private val order: OrderRequest?,
     private val analyticsRepository: AnalyticsRepository,
     private val submitHandler: SubmitHandler<PaymentComponentState<BlikPaymentMethod>>,
 ) : BlikDelegate {
+
+    private val inputData: BlikInputData = BlikInputData()
 
     private val _outputDataFlow = MutableStateFlow(createOutputData())
     override val outputDataFlow: Flow<BlikOutputData> = _outputDataFlow
@@ -58,6 +63,10 @@ internal class StoredBlikDelegate(
     override val uiStateFlow: Flow<PaymentComponentUIState> = submitHandler.uiStateFlow
 
     override val uiEventFlow: Flow<PaymentComponentUIEvent> = submitHandler.uiEventFlow
+
+    init {
+        updateComponentState(outputData)
+    }
 
     override fun initialize(coroutineScope: CoroutineScope) {
         submitHandler.initialize(coroutineScope, componentStateFlow)
@@ -82,7 +91,7 @@ internal class StoredBlikDelegate(
             submitFlow = submitFlow,
             lifecycleOwner = lifecycleOwner,
             coroutineScope = coroutineScope,
-            callback = callback
+            callback = callback,
         )
     }
 
@@ -91,24 +100,39 @@ internal class StoredBlikDelegate(
     }
 
     override fun getPaymentMethodType(): String {
-        return storedPaymentMethod.type ?: PaymentMethodTypes.UNKNOWN
+        return paymentMethod.type ?: PaymentMethodTypes.UNKNOWN
     }
 
     override fun updateInputData(update: BlikInputData.() -> Unit) {
-        Logger.e(TAG, "updateInputData should not be called in StoredBlikDelegate")
+        inputData.update()
+        onInputDataChanged()
     }
 
-    override fun onSubmit() {
-        val state = _componentStateFlow.value
-        submitHandler.onSubmit(state)
+    private fun onInputDataChanged() {
+        Logger.v(TAG, "onInputDataChanged")
+        val outputData = createOutputData()
+        outputDataChanged(outputData)
+        updateComponentState(outputData)
     }
 
-    private fun createOutputData() = BlikOutputData(blikCode = "")
+    private fun createOutputData() = BlikOutputData(inputData.blikCode)
 
-    private fun createComponentState(): PaymentComponentState<BlikPaymentMethod> {
+    private fun outputDataChanged(outputData: BlikOutputData) {
+        _outputDataFlow.tryEmit(outputData)
+    }
+
+    @VisibleForTesting
+    internal fun updateComponentState(outputData: BlikOutputData) {
+        val componentState = createComponentState(outputData)
+        _componentStateFlow.tryEmit(componentState)
+    }
+
+    private fun createComponentState(
+        outputData: BlikOutputData = this.outputData
+    ): PaymentComponentState<BlikPaymentMethod> {
         val paymentMethod = BlikPaymentMethod(
             type = BlikPaymentMethod.PAYMENT_METHOD_TYPE,
-            storedPaymentMethodId = storedPaymentMethod.id
+            blikCode = outputData.blikCodeField.value
         )
 
         val paymentComponentData = PaymentComponentData(
@@ -118,9 +142,14 @@ internal class StoredBlikDelegate(
 
         return PaymentComponentState(
             data = paymentComponentData,
-            isInputValid = true,
+            isInputValid = outputData.isValid,
             isReady = true
         )
+    }
+
+    override fun onSubmit() {
+        val state = _componentStateFlow.value
+        submitHandler.onSubmit(state)
     }
 
     override fun isConfirmationRequired(): Boolean = _viewFlow.value is ButtonComponentViewType
