+++ blik/src/main/java/com/adyen/checkout/blik/internal/ui/model/BlikOutputData.kt
@@ -5,15 +5,16 @@
  *
  * Created by josephj on 4/12/2020.
  */
-package com.adyen.checkout.blik
+package com.adyen.checkout.blik.internal.ui.model
 
+import com.adyen.checkout.blik.R
 import com.adyen.checkout.components.base.OutputData
 import com.adyen.checkout.components.ui.FieldState
 import com.adyen.checkout.components.ui.Validation
 import com.adyen.checkout.core.log.LogUtil
 import com.adyen.checkout.core.log.Logger
 
-class BlikOutputData(blikCode: String) : OutputData {
+internal class BlikOutputData(blikCode: String) : OutputData {
 
     val blikCodeField: FieldState<String>
     override val isValid: Boolean
